Changed code data types to uint16_t

+++ src/huffman.h
@@ -3,8 +3,8 @@
 
 #include "huffman_utils.h"
 
-unsigned int encode(const char *text, unsigned long long int *code);
+unsigned int encode(const char *text, uint16_t *code);
 
-unsigned int decode(const unsigned long long int *code, char *text);
+unsigned int decode(const uint16_t *code, char *text);
 
 #endif
